requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
requests==2.28.2
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.1
